Add support for classes and instance methods
This update creates a new LoxClass, LoxInstance classes and their relevant methods in the interpreter. It also modifies the LoxFunction class to be more compatible with these changes. Moreover, the update adds additional parser and resolver functionality for working with classes, and class types are now differentiated in the resolver.